requirements.txt
absl-py==0.9.0
asgiref==3.2.7
astor==0.8.1
beautifulsoup4==4.9.0
bleach==1.5.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Django==3.0.5
gast==0.3.3
grpcio==1.28.1
gTTS==2.1.1
gTTS-token==1.1.3
h5py==2.10.0
html5lib==0.9999999
idna==2.9
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
pkg-resources==0.0.0
protobuf==3.11.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
